backend/requirements.txt
# API
fastapi[standard]
python-multipart
pydantic

# LangChain
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
tiktoken
pypdf

# Vector store
chromadb

# Database
sqlalchemy[asyncio]
asyncpg

# Auth — JWT verification + password hashing + Google OAuth
python-jose[cryptography]
passlib[bcrypt]
httpx

# Utilities
python-dotenv